requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.32
python-dotenv==1.0.1
APScheduler==3.10.4
google-api-python-client==2.142.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
itsdangerous==2.2.0
reportlab==4.2.5
ics==0.7.2
Jinja2==3.1.4
WTForms==3.1.2
Gunicorn==22.0.0
